\name{Grob-class}
\alias{Grob-class}
\alias{Grob}
\alias{Grob-method}
\alias{Grob,gg-method}
\alias{Grob,gtable-method}
\alias{Grob,trellis-method}
\alias{Grob,lattice-method}
\alias{Grob,GGbio-method}
\alias{GrobList-class}
\alias{ggplotGrob-class}
\alias{latticeGrob-class}
\alias{GrobList}
\title{Grob getter}
\description{
  'Grob' class is a container for 'grob' based object defined with grid
  system. Generic function \code{Grob} gets grob object supported by grid system, and make an instance of
  subclass of class 'Grob'.

  'GrobList' is a container of list of 'Grob' object.
}
\usage{

}
\arguments{
}
\value{
}
\details{
}
\references{
}
\seealso{
}
\examples{
}
\author{Tengfei Yin}

